requirements.dev.txt
Pillow>=10.3.0,<10.4
django>=5.0.2,<5.1
djangorestframework>=3.15.0,<3.16
psycopg2-binary>=2.9.9,<2.10
drf-spectacular>=0.27.2,<0.28
python-keycloak>=4.3.0,<4.4